// src/interfaces/enum.ts
export enum PERMISSION_ENUM {
  SYSTEMADMIN = 1,
  SYSTEMOPS = 2,
  SALEOPS = 3,
  CUSTOMERSERVICE = 4,
}

export const PERMISSION_ALL = [PERMISSION_ENUM.SYSTEMADMIN, PERMISSION_ENUM.CUSTOMERSERVICE, PERMISSION_ENUM.SYSTEMOPS, PERMISSION_ENUM.SALEOPS];

export enum STATUS {
  success = 1,
  fail = 0,
  lock = 2,
}

export enum STATUS_USER {
  ACTIVE = 1,
  UN_ACTIVE = 0,
}

export enum PERMISSION_NAME {
  SYSTEMADMIN = "SYSTEM ADMIN",
  SYSTEMOPS = "SYSTEM OPS",
  SALEOPS = "SALE OPS",
  CUSTOMERSERVICE = "CUSTOMER SERVICE",
}

export enum STATUS_API {
  SUCCESS = "success",
}